Add filtering to attempts by tag
Created a Keyword model separate from Tag to do this, since I want to remove the binding from Tags, which can be changed or deleted at any time. This means that attempts can be filtered independently of the flashcards they are/were associated with and which tags those flashcards are/were associated with as well. So even if you go ahead and burn down all your datasets, you can still look at and filter historic data for how you did (unless you delete the attmpts themselves, of course)