Change in approach for caching - don't really want it on the questions
I'll write more about this elsewhere, but because the question and answer data can be changed at any point, and _usually_ returned after potentially long periods before being immediately modified by the act of attempting it, it makes little sense to be mucking around with caching here. However, it could be useful for the attempts themselves, which are baked-in and immutable.